test: LSP8CompatibilityForERC721 tests both constructor and proxy deployments
